Migrate ContactService to Application layer + delete old Infrastructure services
ContactService moved from Infrastructure (DbContext-dependent) to
Application layer as a coordination service that routes through
IUserService, IUserEmailService, and ICommunicationPreferenceService.

New IUserService methods: GetByEmailOrAlternateAsync, GetContactUsersAsync.
New IUserEmailService/Repository method: FindVerifiedEmailWithUserAsync.
New DTO: UserEmailWithUser for contact conflict checks.

Deleted 4 old Infrastructure services now superseded by Application-layer
versions: CommunicationPreferenceService, UserEmailService,
VolunteerHistoryService, ContactService.
